+++ workflow/src/__init__.py
@@ -1,3 +1,4 @@
+from .constants import TRANSCRIPT_BASED_METHODS, STAINING_BASED_METHODS, SEGMENTATION_METHODS
 from .args import Args
 from .paths import WorkflowPaths
 from .validation import validate_config